Change the default colors for NavigationBar icons and text (title)
Based on this issue https://github.com/shoutem/ui/issues/209 I changed the default colors for NavigationBar icons and text to #222222 (blackish). Heading, title and subtitle style attributes also use this color and it is better to use it in NavigationBar than the previously used #AEAEAE for icons and #666666 for text (title).